test(chat): cover thinking stream in AGUI demo
